refactor lifecycle
